rewrote cache class for storing real data

+++ src/Cache/Cache.php
@@ -7,23 +7,11 @@ use DateTime;
 class Cache
 {
     /**
-     * @var string $cacheConfigFilePath
+     * @var string $cacheFolderPath
      **/
-    private string $cacheFolderPath = '';
+    private string $cacheFolderPath = SERVER_ROOT . '/../cache/';
 
     /**
-     * @var string $cacheConfigFilePath
-     */
-    private string $cacheConfigFilePath = '';
-
-    public function __construct()
-    {
-        $this->cacheFolderPath = SERVER_ROOT . '/../cache/';
-        $this->cacheConfigFilePath = $this->cacheFolderPath . '__cacheConfig.json';
-    }
-
-    /**
-     * function file
      * @param string $identifier
      * @param int $lifeTime
      * @param string $type (public | private)
@@ -87,6 +75,7 @@ class Cache
      * When the max-age = 0 then the http-cache will be removed to force renew cache modified at + lifetime
      * @param int $lifeTime
      * @param int $lastModified
+     * @return array
      */
 
     private function holdCurrentCache(int $lifeTime, int $lastModified): array
@@ -130,14 +119,14 @@ class Cache
     }
 
     /**
-     * function data
      * cache/get cached data
      * @param string $identifier(unique identifier)
      * @param \Closure $closure
      * @param int $lifeTime = 3600
+     * @return mixed
      **/
 
-    public function data(string $identifier, \Closure $closure, int $lifeTime = 3600)
+    public function data(string $identifier, \Closure $closure, int $lifeTime = 3600): mixed
     {
         // krijg informatie terug van cache
         $cacheInfo = $this->getCacheItemInfo($identifier);
@@ -152,23 +141,19 @@ class Cache
         $data = $closure();
 
         // voeg json data toe
-        file_put_contents($this->cacheFolderPath . $cacheInfo->fileName, json_encode($data));
-
-        // put inside config
-        $this->putInsideConfig(['identifier' => $cacheInfo->identifier, 'fileName' => $cacheInfo->fileName, 'lastModified' => time(), 'lifeTime' => $lifeTime]);
+        file_put_contents($this->cacheFolderPath . $cacheInfo->fileName, $lifeTime . ';' . json_encode($data));
 
         // return data
         return $data;
     }
 
     /**
-     * function get
      * get single cache file if exists
      * @param string $identifier(unique identifier)
-     * @return boolean|object
+     * @return bool|object
      **/
 
-    public function get(string $identifier)
+    public function get(string $identifier): bool|object
     {
         // krijg cache infor bij identifier
         $cacheInfo = $this->getCacheItemInfo($identifier);
@@ -178,13 +163,12 @@ class Cache
     }
 
     /**
-     * function remove
      * removes cache file
      * @param string $identifier(unique identifier)
      * @return void
      **/
 
-    public function remove(string $identifier, bool $isFlushed = false)
+    public function remove(string $identifier, bool $isFlushed = false): void
     {
         // krijg cache infor bij identifier
         $cacheInfo = $this->getCacheItemInfo($identifier);
@@ -202,13 +186,12 @@ class Cache
     }
 
     /**
-     * function getCacheItemInfo
      * gets inforamtion cache information by identifier
      * @param string $identifier(unique identifier)
      * @return object
      **/
 
-    private function getCacheItemInfo(string $identifier)
+    private function getCacheItemInfo(string $identifier): object
     {
         // generate hash for identifier
         $identifier = $this->generateFileName($identifier);
@@ -219,137 +202,87 @@ class Cache
             'lastModified' => null,
             'fileName' => $identifier . '.json',
             'identifier' => $identifier,
-            'fileFound' => null
+            'fileFound' => true
         ];
 
         // check if file exists
-        if (!file_exists($this->cacheConfigFilePath)) {
+        if (!file_exists($this->cacheFolderPath . $returnData->fileName)) {
             // update fileFound prop
             $returnData->fileFound = false;
             // return data
             return $returnData;
         }
 
-        // krijg cache config data
-        $cacheConfigData = json_decode(file_get_contents($this->cacheConfigFilePath));
-
-        // kijk of er cache data bestaat
-        if (empty($cacheConfigData) || !is_array($cacheConfigData)) {
-            // update fileFound prop
-            $returnData->fileFound = false;
-            // return data
-            return $returnData;
-        }
-
-        // vind cache item by identifier
-        $cacheItem = array_filter($cacheConfigData, function ($cacheItem) use ($identifier) {
-            return isset($cacheItem->identifier) && $cacheItem->identifier === $identifier;
-        });
-
-        // kijk of er een cacheItem is gevonden
-        if (empty($cacheItem)) {
-            // update fileFound prop
-            $returnData->fileFound = false;
-            // return data
-            return $returnData;
-        }
-
-        // get cacheItem info
-        $cacheItem = $cacheItem[array_key_first($cacheItem)];
-
-        //  kijk of lifetime is verlopen
-        if ($cacheItem->lastModified + $cacheItem->lifeTime <= time()) {
-            // verwijder cache file
-            unlink($this->cacheFolderPath . $cacheItem->fileName);
-            // update return data
+        // check if cache item is expired
+        if (!($cacheItemData = $this->checkCacheItemLifeTime($returnData->fileName))) {
+            // set no file found
             $returnData->fileFound = false;
             // krijg return data van closure
             return $returnData;
         }
 
-        // kijk of bestand gevonden konden worden
-        // update return data
-        $returnData->fileFound = file_exists($this->cacheFolderPath . $cacheItem->fileName);
-        $returnData->data = $returnData->fileFound ? json_decode(file_get_contents($this->cacheFolderPath . $cacheItem->fileName)) : null;
+        // set cache data to return data object
+        $returnData->data = json_decode($cacheItemData);
 
         // return data
         return $returnData;
     }
 
     /**
-     * function flushEndOfLiveTimeCacheItems
+     * This function will check if the lifetime of an specific cache item is not expired
+     * @param string $fileName
+     * @return bool|string
+     */
+
+    public function checkCacheItemLifeTime(string $fileName): bool|string
+    {
+        // get cache item
+        $cacheItemData = explode(';', file_get_contents($this->cacheFolderPath . $fileName), 2);
+
+        // get cache file last modified date
+        $lastModified = filemtime($this->cacheFolderPath . $fileName);
+
+        // get lifetime from cached data
+        $lifeTime = $cacheItemData[0] ?? 0;
+
+        //  kijk of lifetime is verlopen
+        if ($lastModified + $lifeTime <= time()) {
+            // verwijder cache file
+            unlink($this->cacheFolderPath . $fileName);
+            // krijg return data van closure
+            return false;
+        }
+
+        // return true not expired
+        return $cacheItemData[1] ?? null;
+    }
+
+    /**
      * checkes if cacheFiles are still valid and cacheConfig is up to date
      * @return boolean|\DateTime
      **/
 
     public function flushEndOfLiveTimeCacheItems()
     {
-        // kijk of cache config file exists
-        if (!file_exists($this->cacheConfigFilePath)) {
-            return false;
-        }
-
-        // decode json to array/object
-        $cacheItems = json_decode(file_get_contents($this->cacheConfigFilePath));
-
         // loop through alle cache items and check if the lifetime didn't passed
-        foreach ($cacheItems as $key => $cacheItem) {
-            // check if lifeTime is passed
-            if (!isset($cacheItem->lastModified, $cacheItem->lifeTime, $cacheItem->identifier) || !file_exists($this->cacheFolderPath . $cacheItem->fileName) || $cacheItem->lastModified + $cacheItem->lifeTime <= time()) {
-                // verwijder cache item van array
-                unset($cacheItems[$key]);
-                // check if file exists then remove cache file
-                $this->remove($cacheItem->identifier, true);
-            }
-        }
-
         // wrong files to filter out
-        $wrongFiles = ['..', '.', '.DS_Store', '__cacheConfig.json'];
+        $wrongFiles = ['..', '.'];
 
-        // get all files accept '..' and '.'
-        $files = array_filter(scandir(SERVER_ROOT . '/../cache'), fn ($file) => !in_array($file, $wrongFiles));
-
-        // get all identifiers from cacheConfig items
-        $identifiers = array_column($cacheItems, 'identifier');
-
-        // loop trough all files
-        foreach ($files as $file) {
-            // kijk of file niet in cache config identifiers array staat
-            if (!in_array(str_replace('.json', '', $file), $identifiers)) {
-                // verwijder bestand
-                unlink($this->cacheFolderPath . $file);
+        // loop trough all files and delete them
+        foreach (scandir(SERVER_ROOT . '/../cache') as $file) {
+            // let only valid files pass
+            if (in_array($file, $wrongFiles)) {
+                continue;
             }
+            // remove cache item if the lifetime is expired
+            $this->checkCacheItemLifeTime($file);
         }
-
-        // put new content to cache config file
-        $this->putInsideConfig($cacheItems, false);
 
         // return last updated date
         return (new \DateTime())->format('Y-m-d H:i:s');
     }
 
     /**
-     * function putInsideConfig
-     * updates cache config based on cache data files
-     * @return void
-     **/
-
-    private function putInsideConfig($data, bool $mergeOldData = true): void
-    {
-        // behoud oude email
-        $oldData = [];
-
-        // kijk of cacheConfig file bestaat
-        if ($mergeOldData && file_exists($this->cacheConfigFilePath)) {
-            $oldData = json_decode(file_get_contents($this->cacheConfigFilePath)) ?? [];
-        }
-
-        // insert new data
-        file_put_contents($this->cacheConfigFilePath, json_encode($mergeOldData ? [$data, ...$oldData] : $data));
-    }
-
-    /**
-     * function generateFileName
      * makes hashed string to make filenames more random/secure
      * @param string $identifier
      * @return string
@@ -361,7 +294,6 @@ class Cache
     }
 
     /**
-     * function flush
      * This method will delete all cache files
      * @return self
      */
@@ -376,11 +308,13 @@ class Cache
         // wrong files to filter out
         $wrongFiles = ['..', '.'];
 
-        // get all files accept '..' and '.'
-        $files = array_filter(scandir(SERVER_ROOT . '/../cache'), fn ($file) => !in_array($file, $wrongFiles));
-
         // loop trough all files and delete them
-        foreach ($files as $file) {
+        foreach (scandir(SERVER_ROOT . '/../cache') as $file) {
+            // let only valid files pass
+            if (in_array($file, $wrongFiles)) {
+                continue;
+            }
+            // delete cache file
             unlink($this->cacheFolderPath . $file);
         }
